+++ src/main/java/uk/gov/pay/connector/charge/validation/telephone/CardTypeValidator.java
@@ -1,0 +1,28 @@
+package uk.gov.pay.connector.charge.validation.telephone;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+import java.util.HashSet;
+
+public class CardTypeValidator implements ConstraintValidator<ValidCardType, String> {
+    private final static HashSet<String> CARD_TYPES = new HashSet<>();
+
+    static {
+        CARD_TYPES.add("master-card");
+        CARD_TYPES.add("visa");
+        CARD_TYPES.add("maestro");
+        CARD_TYPES.add("diners-club");
+        CARD_TYPES.add("american-express");
+    }
+
+    @Override
+    public boolean isValid(String value, ConstraintValidatorContext context) {
+
+        if (value == null) {
+            return false;
+        }
+
+        return CARD_TYPES.contains(value);
+
+    }
+}
