+++ src/main/java/uk/gov/pay/connector/charge/validation/telephone/ValidCardType.java
@@ -1,6 +1,4 @@
-package uk.gov.pay.connector.charge.validation;
-
-import uk.gov.pay.connector.charge.validation.telephone.CardExpiryValidator;
+package uk.gov.pay.connector.charge.validation.telephone;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -15,16 +13,15 @@ import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
 import static java.lang.annotation.ElementType.TYPE_USE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@Target({FIELD, METHOD, PARAMETER, ANNOTATION_TYPE, TYPE_USE})
+@Target({ FIELD, METHOD, PARAMETER, ANNOTATION_TYPE, TYPE_USE })
 @Retention(RUNTIME)
-@Constraint(validatedBy = CardExpiryValidator.class)
+@Constraint(validatedBy = CardTypeValidator.class)
 @Documented
-public @interface ValidCardExpiryDate {
-
-    String message() default "Must be MM/YY";
-
-    Class<?>[] groups() default{};
-
-    Class<? extends Payload>[] payload() default{};
+public @interface ValidCardType {
     
+    String message() default "Must be either master-card, visa, maestro, diners-club or american-express";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
 }
